🐛 Fix fetch error
